Convert entity annotations mapping to xml mapping
