+++ src/Afup/Model/Group.php
@@ -1,0 +1,141 @@
+<?php
+
+namespace Afup\Model;
+
+use Doctrine\ORM\Mapping as ORM;
+
+use FOS\UserBundle\Model\Group as BaseGroup;
+
+/**
+ * Group
+ */
+class Group extends BaseGroup
+{
+    /**
+     * @var string
+     */
+    private $name;
+
+    /**
+     * @var integer
+     */
+    private $id;
+
+    /**
+     * @var \Doctrine\Common\Collections\Collection
+     */
+    private $roles;
+
+    /**
+     * @var \Doctrine\Common\Collections\Collection
+     */
+    private $members;
+
+    /**
+     * Constructor
+     */
+    public function __construct()
+    {
+        $this->roles = new \Doctrine\Common\Collections\ArrayCollection();
+        $this->members = new \Doctrine\Common\Collections\ArrayCollection();
+    }
+
+    /**
+     * Set name
+     *
+     * @param string $name
+     * @return Group
+     */
+    public function setName($name)
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * Get name
+     *
+     * @return string 
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * Get id
+     *
+     * @return integer 
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * Add roles
+     *
+     * @param \Afup\Model\GroupRole $roles
+     * @return Group
+     */
+    public function addRole(\Afup\Model\GroupRole $roles)
+    {
+        $this->roles[] = $roles;
+
+        return $this;
+    }
+
+    /**
+     * Remove roles
+     *
+     * @param \Afup\Model\GroupRole $roles
+     */
+    public function removeRole(\Afup\Model\GroupRole $roles)
+    {
+        $this->roles->removeElement($roles);
+    }
+
+    /**
+     * Get roles
+     *
+     * @return \Doctrine\Common\Collections\Collection 
+     */
+    public function getRoles()
+    {
+        return $this->roles;
+    }
+
+    /**
+     * Add member
+     *
+     * @param \Afup\Model\Member $member
+     * @return Group
+     */
+    public function addMember(\Afup\Model\Member $member)
+    {
+        $this->members[] = $member;
+
+        return $this;
+    }
+
+    /**
+     * Remove member
+     *
+     * @param \Afup\Model\Member $member
+     */
+    public function removeMember(\Afup\Model\Member $member)
+    {
+        $this->members->removeElement($member);
+    }
+
+    /**
+     * Get member
+     *
+     * @return \Doctrine\Common\Collections\Collection 
+     */
+    public function getMember()
+    {
+        return $this->members;
+    }
+}
