+++ src/Afup/Model/Mapping/Group.orm.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+  <entity name="Afup\Model\Group" table="group">
+    <id name="id" type="integer" column="id">
+      <generator strategy="IDENTITY"/>
+    </id>
+    <field name="name" type="string" column="name" precision="0" scale="0" nullable="false"/>
+    <one-to-many field="roles" target-entity="Afup\Model\GroupRole" mapped-by="group" orphan-removal="true">
+      <cascade>
+        <cascade-remove/>
+      </cascade>
+    </one-to-many>
+    <many-to-many field="members" target-entity="Afup\Model\Member" mapped-by="groups"/>
+  </entity>
+</doctrine-mapping>
